+++ docs/changelog.rst
@@ -1,9 +1,15 @@
 Change history
 ==============
 
+1.2.5a2 (2018-10-30)
+
+* Add ModelSchema(Auto generate load and dump func for EnumField).
+* Add logging config file.
+* Fix use_kwargs with fileds.missing=None.
+
 1.2.5a1 (2018-10-25)
 
-* Added EnumExt implementation
+* Add EnumExt implementation.
 
 1.2.5a0 (2018-10-22)